♻ 인터셉터에서는 인증만을 담당하고 권한체크는 표현 계층에서 수행하도록 수정
- 인터셉터에서 인증과 권한을 모두 체크하여 view 를 분기하던 어색한 구조를 제거한다.
- AuthController 는 로그인 관련 여부만 검사하고 Controller 단에서 세션 정보를 사용하여 권한을 검사한다.

+++ src/main/java/com/kakao/cafe/controller/ArticleController.java
@@ -1,13 +1,21 @@
 package com.kakao.cafe.controller;
 
+import com.kakao.cafe.domain.Article;
 import com.kakao.cafe.dto.ModifiedArticleParam;
 import com.kakao.cafe.dto.NewArticleParam;
 import com.kakao.cafe.exception.article.ArticleDomainException;
+import com.kakao.cafe.exception.common.AccessRestrictionException;
+import com.kakao.cafe.exception.common.CommonException;
 import com.kakao.cafe.service.ArticleService;
+import com.kakao.cafe.session.SessionUser;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
+
+import static com.kakao.cafe.message.ArticleDomainMessage.DELETE_ACCESS_RESTRICTION_MESSAGE;
+import static com.kakao.cafe.message.ArticleDomainMessage.MODIFY_ACCESS_RESTRICTION_MESSAGE;
 
 @Controller
 @RequestMapping("/articles")
@@ -33,20 +41,47 @@ public class ArticleController {
     }
 
     @GetMapping("/{id}/form")
-    public ModelAndView goUpdateForm(@PathVariable int id, ModelAndView mav) {
+    public ModelAndView goUpdateForm(@PathVariable int id,
+                                     @SessionAttribute("sessionUser") SessionUser sessionUser,
+                                     ModelAndView mav) {
+
+        Article article = articleService.search(id);
+        String userId = article.getWriter();
+
+        if (!sessionUser.ownerOf(userId)) {
+            throw new AccessRestrictionException(HttpStatus.FORBIDDEN, "error/4xx", MODIFY_ACCESS_RESTRICTION_MESSAGE);
+        }
+
         mav.setViewName("qna/updateForm");
         mav.addObject("article", articleService.search(id));
         return mav;
     }
 
     @PutMapping("/{id}")
-    public String modifyArticle(ModifiedArticleParam modifiedArticleParam) {
+    public String modifyArticle(ModifiedArticleParam modifiedArticleParam,
+                                @SessionAttribute("sessionUser") SessionUser sessionUser) {
+
+        String userId = modifiedArticleParam.getWriter();
+
+        if (!sessionUser.ownerOf(userId)) {
+            throw new AccessRestrictionException(HttpStatus.FORBIDDEN, "error/4xx", MODIFY_ACCESS_RESTRICTION_MESSAGE);
+        }
+
         articleService.update(modifiedArticleParam);
         return "redirect:/";
     }
 
     @DeleteMapping("/{id}")
-    public String removeArticle(@PathVariable int id) {
+    public String removeArticle(@PathVariable int id,
+                                @SessionAttribute("sessionUser") SessionUser sessionUser) {
+
+        Article article = articleService.search(id);
+        String userId = article.getWriter();
+
+        if (!sessionUser.ownerOf(userId)) {
+            throw new AccessRestrictionException(HttpStatus.FORBIDDEN, "error/4xx", DELETE_ACCESS_RESTRICTION_MESSAGE);
+        }
+
         articleService.remove(id);
         return "redirect:/";
     }
@@ -54,5 +89,13 @@ public class ArticleController {
     @ExceptionHandler({ArticleDomainException.class})
     private ResponseEntity<String> except(ArticleDomainException ex) {
         return new ResponseEntity<>(ex.getMessage(), ex.getHttpStatus());
+    }
+
+    @ExceptionHandler({CommonException.class})
+    private ModelAndView goErrorPage(CommonException ex) {
+        ModelAndView mav = new ModelAndView(ex.getViewName());
+        mav.setStatus(ex.getHttpStatus());
+        mav.addObject("message", ex.getMessage());
+        return mav;
     }
 }
